Add intent-filter for ACTION_GET_CONTENT
This is an experimental support to be able to teamfood the new
PhotoPicker UI. The intent-filter will be removed in production build.

Added activity-alias that can receive ACTION_GET_CONTENT requests. This
activity will be disabled by default. It can be enabled by enabling
activity through command line.

By default, we return redactedUri for non-picker intent actions because
we have to support file path/query support for most of the apps. Sending
the picker uri may break lot of apps. Added a system property to tune
picker uri Vs redacted uri.

Bug: 168001592
Test: Verified that MediaStore app is displayed in documentsui and we
are able to give read access to picked images.

Change-Id: Iaa9e2dfb6c4c49f948c862dd22858baed0c74a69