Add a photo picker database
This database will store a cache of merged/deduped cloud and local
media items containing a subset of columns relevant to the photo
picker.

It is important that these information is in a separate db and not in
the external db for the following reasons:
1. Avoiding performance regressions in existing MediaProvider code
resulting from potentially large data from the cloud
2. The characteristics of this db are very different from those of the
external db, for example, more tolerance for slow writes and db
corruption since the data can be easily re-synced from the local and
cloud providers in the background. In future, this can allow us to
tune db parameters like journaling mode without impacting the rest of
the MediaProvider.

This change introduces the db with one table, 'media' to store media
metadata. It also introduces some helpful indices on important
columns. Additionally, we skip any upgrade or downgrade code for now
since we are just introducing the db.

In future cls, we'll introduce the 'albums' and 'media_albums' table
to store album metadata and media to album item mappings respectively.

Test: atest DatabaseHelperTest
Bug: 190713331
Change-Id: Ice2225ce428bc34c6f82c88a8d0d5218699fb3aa