Undeprecate DATA column
Android 10 did not support file path operations for apps in
scoped storage. But Android 11 and above do support file path
operations for such apps. So DATA column is useful again though
it's read-only for apps that target R and later.

Undeprecate and fix javadoc to reflect the current level of support.

BUG: 153562415
Test: m
Change-Id: I560d054a8be074ccda68fd0fed55da98d656db27