Support updating the picker db
Added a helper class to facilitate adding/removing cloud and local
items for the picker db.

This helper handles the details of inserting, updating or deleting
rows with synced items. It also allows querying for the merged and
deduped (visible) items. Note that for associated local and
cloud+local items, the local item is always preferred.

Bug: 190713331
Test: atest PickerDbFacadeForPickerTest

Change-Id: Ia4c87ca68578c7bd6c781b26d95aaf02f440098e