Stabilise test by removing complicated checks
testNoOpScan_NoMediaDirs primarily tests that nomedia dirs should only
be scanned once. This is a test for an issue which was over-using CPU
resources and over-heating the device on scans. As nomedia directories
can be very large, nomedia dirs should not be scanned again if files are
added/modified/removed.

They should only be scanned again if their location is changed (or they
are renamed). This second check is not the main primary motive for this
test, so we can remove it for stabilising the test.

Bug: 188981504
Bug: 191845116
Test: atest ModernMediaScannerTest

Change-Id: Ifef6323f8f6abe7c9375ebe843ce91ec9e2afb15
Merged-In: Ifef6323f8f6abe7c9375ebe843ce91ec9e2afb15
(cherry picked from commit 45f36d089a25495737ee88f1ebcf5e96abc0eb15)