Relax local_id unique constraint for the picker db
Previously, the local_id column was required to be unique. This
introduced two limitations:

1. cloud+local media items synced from the cloud provider and
local-only items synced from the local provider have to be stored in
the same row to satisfy the local_id uniqueness. This made it
difficult to represent a local and cloud+local item sharing a local_id
with different metadata. Specifically, we want to prioritise the
local-only item, but if it gets deleted, we'll need to update the row
with the latest metadata from the cloud provider.
2. Multiple cloud+local items mapping to the same local_id had to be
stored in the same row (effectively hiding one of the cloud+local
items). This means, there was no way to represent the following in the
db without hiding one of the cloud_ids:
|local_id|cloud_id|
|1       |a       |
|1       |b       |

To workaround these issues, we introduce a new 'is_visible' column
and discard the 'is_local_verified' column. Additionally, we relax the
local_id unique constraint to UNIQUE(local_id, is_visible). The
UNIQUE(cloud_id) constraint is left the same.

With this change, each local and cloud item will have its own
row. Only one of them can have is_visible=1 while the others will be
is_visible=NULL. This allows us to query for only visible items while
keeping the other local_ids as hidden items in the db. Note that NULLs
are distinct in terms of unique constraints, so we can have multiple
rows sharing the same local_id as long as their is_visible=NULL.

This change has the folllowing benefits:
1. Significantly reduces the code complexity during db syncs while
still relying on simple SQL statements for improved performance
2. Fixes the problems mentioned above
3. Paves the way to allow multiple cloud providers sync into the
db. We can simply add an 'authority' column for each row and local_ids
will still be deduped across cloud providers

Bug: 190713331
Test: atest PickerDatabaseHelperTest
Change-Id: Ib9d1e63ac10ec150ae0cedd889daf527b7e97f4f