Change UUID after database upgrade
Change the UUID when we are upgrading to new version. This is useful
when we revert back to the same version(because of upgrade + downgrade)
with major reset in database. Changing the UUID during upgrade will
ensure MediaStore#getVersion is different after the reset

Bug: 192102594
Test: atest
com.android.providers.media.DatabaseHelperTest#testUpgradeAndDowngradeChangeUUID

Change-Id: I7936fff031ec14b0b35ceebc0fb0b997448ba01c