Fix fromExternal() for legacy app clone solutions.
Some legacy app clone solutions redirect clone user native file access
to the owner user; these calls have a LocalCallingIdentity that is
created by fromExternal(), which assigns the user-id based on the uid of
the calling user. This user is subsequently rejected access to the
volume, because the calling user does not match the user owning the
volume.

Fix this by checking with the UserCache whether this user shares media
with its parent; if not, we know this device is not using the new app
clone design, and we use the owner user instead. This is similar to what
we already do in fromBinder().

Bug: 193590928
Bug: 193852033
Test: TEST_MAPPING

Change-Id: I388fcb9d320b702017657c25c112fec7eec63b00