Redefine XATTR_UUID to XATTR_UUID_V2
We need a one time reset of UUID for MediaProvider database. This is
necessary to ensure we reset UUID for all devices with new code even
when there is no MediaProvider database upgrade required.
This is usually achieved by doing a database version bump and
resetting the UUID in the upgrade routine. The version bump can only be
done on the VERSION_LATEST or if there is a gap in version change, we
could use that. Since there are no version gaps, we can't change the
version, hence we redefine XATTR_UUID to XATTR_UUID_V2 which allows us
to do one time reset of XATTR_UUID

Bug: 192102594
Test: atest
com.android.providers.media.DatabaseHelperTest#testVerifyXATTR_UUID_V2

Change-Id: I4690c21e734465ad481103639c03183471b6ca32