Remove unnecessary permission checks in CloudMediaProvider
These checks prevent us from adding an exported=false
CloudMediaProvider implemented by the MediaProvider. This is because,
the client side code enforces exported=true and even though the
provider is in our process and we have access to it, the client side
checks throw an exception

Additionally, the checks are redundant since the MediaProvider will
check the same before connecting to the provider.

Minor fixes:
1. Exposed some useful constants via the CloudMediaProviderContract
2. Changed URI exact matching pattern to only include id without
version. Since ids will be stable, this is sufficient.

Test: m
Bug: 190713331
Change-Id: Ic459be958206b3be7df3151f7564a3be54d81e29