Fix fd leak while getting original media format fd
To avoid transcoding videos when an app uses the platform media APIs,
we convert a transcoding fd to a non-transcoding fd. To do
this, we expose an API that apps can use to pass in an fd to be
converted: MediaStore#getOriginalMediaFormatFileDescriptor. This API
leaked fds twice in the MediaProvider:

1. The input fd was passed as a Parcelable in a Bundle and was never
closed. A simple fix was to wrap the input fd in a try auto-close
block so it gets closed after use
2. The output fd was also passed as a Parcelable in a Bundle and was
never closed. In this case though, we can't simply wrap the code in a
try auto-close block because we'd end up closing the fd before it
leaves the MediaProvider process. The fix here relies on the
PARCELABLE_WRITE_RETURN_VALUE flag behavior that auto-closes fds after
they are returned from ContentProviders. This required re-implementing the
MediaStore API over the ContentResolver#openTypedAssetFileDescriptor
instead

See I978257bbc4a8f6813b6e6a5ce22124257204f432 for another 2 fd leak
fix in the client app

Test: atest TranscodeTest
Test: Manual with 'adb shell lsof | grep <filename>'
Bug: 194828489
Bug: 179804072

Change-Id: I0124ec8fbd0471237e99bab321f903544f8fe1f8
Merged-In: I0124ec8fbd0471237e99bab321f903544f8fe1f8