Introduce a cache for work profile owner apps
It is wasteful to check all installed packages for each cross-profile
file access. So a new cache has been introduced to reduce the time to
see if the userId is for work profile.

When a package which is registered as the profile owner for the user is
removed, the cache is invalidated.

Bug: 192041648
Test: atest MediaProviderTests
Change-Id: I08a6a808ce918977b059c759cb5680eccbbddec8