Define MANAGE_CLOUD_MEDIA_PROVIDERS permission in MediaProvider
This permission will be used to protect CloudMediaProviders. Because
the permission is defined in the MediaProvider mainline module as
signature only, the MediaProvider will have exclusive access to the
permission and it will be available on R and S devices where the
module is backported to.

We also grant MediaProvider itself the permission

Bug: 195008823
CTS-Coverage-Bug: 193668830
Test: adb shell dumpsys package
com.google.android.providers.media.module | grep
manage_cloud_media_providers

Change-Id: Icd1657d0f91366a28f5cced52329d9b9289c3877