Add local and cloud provider awareness to the PickerDbFacade
When displaying results from the picker db in the UI, we need to know
the cloud provider contained in the cursor. To avoid racing between
the Cursor result of the PickerDbFacade#query and the
PickerSyncController#setCloudProvider which might change the cloud
provider while querying the db, we made the following design
improvements:

1. The facade returns the local_provider and cloud_provider as
part of it's Cursor result extras.
2. The facade is instantiated with the local_provider while
the cloud_provider can be set by the controller
3. Simplified the facade interface to accept 'authority' and decide if
the operation is for local or cloud
4. Changing the current cloud_provider is a two-step process:
a. controller always sets the facades cloud_provider to null, thus
clearing the cloud_provider from subsequent queries
b. After every sync, the controller sets the facades cloud_provider to
the latest cloud_provider thus making the latest cloud_provider
available in subseqent queries

Bug: 195009139
Test: atest PickerDbFacadeTest
Test: atest PickerSyncControllerTest

Change-Id: Ib68f925b54534de901a981c94338db69f244fa6c