Fix deleted_media notification
If a media file is inserted as pending and then updated to
non-pending, we should always notify the sync controller.

Previously, we only notified the sync controller if we successfully
removed an item from deleted_media but the file might have never
existed in deleted_media if it was just inserted as pending.

Now, we notify regardless of the results of the deleted_media table
operation success

Test: atest ExternalDbFacadeTest
Bug: 195009139
Change-Id: I33612f223e186d2e9140537d3a2fb768536175b0