Change is_favourite filter to a media column
To retrieve favourited items from cloud media providers, the previous
API used a query filter, EXTRA_FILTER_IS_FAVOURITE. With this API, it
was difficult to efficiently maintain an up to date cache of the
merged local and cloud favourite category.

Now, we retrieve favourited items with an IS_FAVOURITE column on the
media item.

Bug: 195008671
Test: atest PickerDbFacadeTest
CTS-Coverage-Bug: 193668830
Change-Id: I66f150df9aaad3c08dc59ede8eddb89b8a5d498f