Allow FUSE file opening in Java
The MediaProvider can override the FUSE_OPEN behavior by providing an
fd from the Java layer. If this happens, the FuseDaemon.cpp skips the
open(2) call.

This is the first in a stack of cls to service FUSE open requests via
a cloud provider. This change is a no-op if the MediaProvider doesn't
pass an fd from Java

Test: Manual, file opens work
Bug: 195009143
Change-Id: I649403be2ea9993f21438bc4d711dbda25979662