Modify API behavior for PhotoPicker multi-selection flow
In an attempt to reduce implicit parameters for PhotoPicker, only
accept explicit values specified by the developer.
This helps developers to make explicit decisions on the number of items
they can handle. For example: An app should not crash if the user
selects 70 images but the app can only handle 10 at a time.

Also, add an API to indicate max limit of items that can be selected in
PhotoPicker. This is essential for apps that can handle as many images
to be selected as are allowed by PhotoPicker.

Bug: 185228105
Test: atest PickerViewModelTest
Test: atest PhotoPickerTest
Change-Id: I3ec396a35905edd9d8c5bbc7530630325bf123fd