Fix the same name issue when we extend the expired items
If there are expired items has same name in the same folder, we use
the same extended time to extend the expiration time of them. When we
update the later item in the database, it throws Exception. To fix it,
add the index of extended items into the extended time to avoid the
issue.

Test: atest IdleServiceTest
Bug: 199469244
Fix: 187650056
Change-Id: I54ceb462ab6bea5043d6f9005ddff73e795d7f3f
Merged-In: I54ceb462ab6bea5043d6f9005ddff73e795d7f3f