Fix transcoding lookup result error
I8dd3d2503cc1f3f38b4cfde8730ab82e843b5972 introduced a bug where we
always set FLAG_TRANSFORM_TRANSCODING as long as transcoding was
supported, regardless of if it was actually required or not.

This caused files that shouldn't require transcoding (and hence no
ioPath) to incorrectly attempt transcoding to an empty ioPath.

Test: atest TranscodeTest
Bug: 195009143
Change-Id: I6f77776cc5189fd1e61803ec744a966bb1ec4d0d