Sort based on _ID for rows with same date_taken or date_modified value
This will allow us to be able to predict the order of the images in main
grid for items that have same date_taken or date_modified value when
date_taken is not set.

Being able to predict the order will assist us in preserving the same
order of items in the multi select preview.

Bug: 185801129
Test: atest
com.android.providers.media.photopicker.ItemsProviderTest#testGetItems_sortOrder

Change-Id: I5a3be3e68131e44d0060bce5fe18d0c19d0d0830