Don't use the path when comparing MediaVolume objects.
When we receive an unmount event for a StorageVolume, its path is
always 'null'; therefore, the MediaVolume we construct from that volume
also has a 'null' path, and it fails to compare against an existing
MediaVolume with a path that we may have cached internally.

Since a particular MediaVolume should anyway never be mounted in two
different places, we don't need to use the path for equals(), anyway, so
remove it.

Bug: 201253905
Test: TEST_MAPPING
Change-Id: I91cd4ca3ac305881bc9f5b71f0f1c833d6b2262b
(cherry picked from commit 9dece9a997d16955db443d45a1b78cbd9c5060eb)