Support album and mimeType filtering for the local provider
To enable the picker db flag for the album grid, the providers (local
and cloud) need to support the following:
1. media: filtering by album and mimeType
2. albums: filtering by mimeType

We added this support to the ExternalDbFacade used by the local
provider (PhotoPickerProvider).

We also fixed the local provider open methods to run with
MediaProvider's calling identity. This is safe because apps can never
talk to the local (or cloud) providers directly and the MediaProvider
handles permission checks for open requests on the picker URI before
forwarding to the respective local/cloud providers.

Test: atest ExternalDbFacadeTest
Bug: 195009148

Change-Id: Ib06a4aab161dda4b3121bf8b4e2154288a4a2cdf