FuseDaemon uses FUSE-BPF
Enable the FUSE BPF feature in the FUSE daemon.
With this change the FuseDaemon will fetch the FUSE BPF file descriptor
and install it into the data and obb folders. The contents of data and
obb would instead be handled in full pass-through mode by passing only
the file descriptor of the backing folders. Apps would not be able to
access the contents of the private folders of other apps as ACL controls
are still enabled.

FUSE BPF is limited to the primary volume, as the performance bottleneck
of current external storage devices are the devices themselves and not
the FUSE architecture.

Bug: 202785178
Test: CtsScopedStorageHostTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I1011202f81f210b866e64aaee5063e9afd06d80f