Workaround to expose transcode support via stat(2)
We reused the existing st_nlink stat field to mark files
supporting transcoding (regardless of caller). If a file can be
transcoded, the st_nlink field returns '2' instead of the default of
'1'.

This is an ugly hack to share information about transcode support for a
file from the MediaProvider to the MTP stack, but we converged on this
for the following reasons:
1. It is safe because hardlinks are not supported on FAT filesystems
which FUSE is emulating, hence, apps shouldn't be relying on the
st_nlink field
2. It is the most performant way to share this info (avoid binder
calls).
2. It is the smallest and least risky change to make given the urgency
of the fix during the Android release

This change will be complemented with an MTP change to double the
returned MTP object size metadata if the MTP 'Convert to AVC option' is
enabled.

Test: adb shell stat /sdcard/DCIM/Camera/<transcodde supported video>
| grep Links. Verify 'Links: 2'
Bug: 184117074
Bug: 190422448

Change-Id: I4e7f5d7362a5b293c49a7104f7670d796dad0976
(cherry picked from commit 75cd63f4e168febf21bec02d3d51aa132b2d86a5)
Merged-In:I4e7f5d7362a5b293c49a7104f7670d796dad0976