Fix picker URI open
When opening picker URIs, we do the following:
1. Fetch the corresponding picker file path from the picker db
2. Open the file path from #1 via FUSE. This triggers the logic from
the FUSE daemon to create and open a synthetic file path backed by the
corresponding cloud (or local) provider.

Previously, we were opening the lower fs path in #2 instead of first
converting it to a FUSE file path. This returned an ENOENT since the
FUSE synthetic file path logic wasn't trigerred.

Test: adb shell setprop sys.photopicker.pickerdb.enabled true
&& atest PhotoPickerTest
Bug: 195009143

Change-Id: Ifa7381fedff8f3847ba7d4f16a0d418597bb0faa