Support uncached paths RRO
A new Runtime Resource Overlay config_supported_uncached_relative_paths
specifies the paths that should always be accessed without caching.

The config can be used to avoid cache synchronization issues when a
specific path can be accessed from its lower filesystem.

As the name suggests, each item must be specified as a path relative to
the mount point's root. Also, it's assumed to point to a directory and
must end with a slash '/'.

By default, the config array is empty.

Bug: 208964465
Test: Add a relative path to config_supported_uncached_relative_paths
      and confirm that files under the path is accessed without caching.
Test: atest com.android.providers.media.util.StringUtilsTest

Change-Id: If13a3a108cae4c2893e57d9a763c6fd51d037509