Silence clang tidy performance-no-int-to-ptr warnings
The `uint64_t fuse_file_info::fh` is a file handle identifier the type
of which depends on the FUSE daemon implementation, thus castings are
always required. FUSE has other examples of numeric types that are often
converted to pointers.
Clang tidy throws warnings for the casting of these integers to
pointers, and the noisy output may hide other iportant warning messages.

Disable the performance-no-int-to-ptr tidy check to reduce the warnings.

Bug: 202785178
Test: m
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I03debacc7b0a2ad395087ce9cce623a1395bbea3