Refactor PickerDbFacade to support PickerMediaColumns
Previously, the PickerDbFacade only allowed querying with *all* the
CloudMediaProviderContract#MediaColumns. This is not suitable for apps
querying with the newly introduced MediaStore#PickerMediaColumns,
especially if they want control of the projections used.

Now, we refactored PickerDbFacade query methods with the following
changes:

1. s/queryMedia/queryMediaWithCmpContract/. This returns a Cursor
with *all* the CMP columns as before
2. s/queryMediaId/queryMediaIdWithMediaStoreContract/. This now
returns a Cursor with the MediaStore columns given a projection input

Test: atest PickerDbFacadeTest
Test: atest PickerSyncControllerTest
Test: atest CtsPhotoPickerTest
Bug: 201385900

Change-Id: I6055d6c526028f3c698919eadb40ca794d8dfae7
Merged-In: I6055d6c526028f3c698919eadb40ca794d8dfae7