Throw IllegalArgumentException for getContentResolverForUserId
Throwing FileNotFoundException was not the best choice, as all callers
of getContentResolverForUserId would have to handle it in their own way.
IllegalArgumentException makes sense as we are unable to find a content
resolver corresponding to the user-id embedded in the picker uri.
Make sure we are consistent in error handling for Picker uris with
MediaProvider

Bug: 202388743
Test: atest PickerUriResolverTest
Change-Id: I0d28f0f653d8a3a2fe3e23442f073e58a64c149a