Fix default folder creation on public volumes.
A recent change switched to using volume IDs as a key for whether we'd
previously created default folders on a volume. This is fine for
primary volumes, because the identifier will be unique per-user.

But for public volumes, the volume identifier might not change, even
when the disk is formatted again; in those cases, use the volume name
instead, which *will* change when formatted.

Bug: 201195355
Bug: 209664271
Test: com.android.providers.media.PublicVolumeTest

Change-Id: Ida4d270247c1442393b7704b39bc572652e4e41f