Correctly handle stable dependency to content providers
This is essential as PhotoPicker should not crash if remote content
provider crashes. For example the following flow will crash PhotoPicker
without this fix:
1. Open PhotoPicker from personal profile app
2. Click on "Switch to Work" button
3. PhotoPicker loads data from user 10 MediaProvider. If a stable
connection is made, then we now have a dependant connection on user 10
MediaProvider.
4. User goes back to Personal Profile
5. Work Profile is switched off
6. PhotoPicker crashes even though we are not in work profile mode. This
is because we created a dependency on a process that is dying now.

Fix this by closing asset file descriptor correctly.

Bug: 191158884
Test: manual
Change-Id: Ie928a44bcf80aea389e17b2f7536797cd0b0d082