Removed @ExportedSince annotation from is_recording.
This annotation was added in M-2021-12, causing the is_recording column
to become invisible on R devices. However, when those R devices update
to S, the views are not recreated to include is_recording again, because
the schema doesn't change.

To fix this, up the schema version and remove the annotation; for
devices already affected by this, upping the schema version will fix the
problem. For devices still on R + M-2021-12, removing the annotation
makes sure that the column becomes visible again, making sure that the
later OTA to S won't cause any issues.

Bug: 211068960
Test: TEST_MAPPING
Change-Id: I11a712ca09560246f025e1a98a9ba67c7e06d28f