Remove unnecessary locks from UserIdManager class
All the getters and setters for UserIdManager class are called from the
main thread. Locks were required previously when broadcast listener was
in UserIdManager class. So, the values of UserIdManager class could be
altered by broadcast listeners in UserIdManager and could be accessed by
the main UI thread concurrently. This is why we previously required
locks.
The broadcast listener is now moved to the TabFragment/PreviewFragment as
there are more broadcasts that PhotoPicker needs to listen to
and handle on the main UI thread (for example if work profile is turned
off or on while the user is in personal or work profile). Moving
broadcast listener to the fragments also help PhotoPicker to not
repeatedly check if the work MediaProvider is available or not.

Calling locked methods from a UI thread may lead to ANRs.

Bug: 199823393
Test: atest UserIdManagerTest
Change-Id: I505f1a3110f55a494c1091bbb5f51f3176a27465