Clear sync state when clearing the picker db
On create/upgrade/downgrade of the picker db, the db is flushed
however the sync state stored in SharedPreferences is left
intact. A subsequent sync will be an incremental sync instead of a
full sync, causing the PhotoPicker to show partial data in the UI.

Now, we separate the user prefs from sync prefs in the
SharedPreferences, this allows us to easily clear all sync prefs while
keeping the user prefs, e.g. the currently enabled cloud provider
intact.

Test: atest PickerSyncControllerTest
Bug: 213144765
Change-Id: Iaf065b258d979ba2db24393aade7868b7482d437