Fix MP row id reuse issue
As part of this CL, we store the next row id to be used per database in
case of database loss/rollback. On every open() call to Database, we
identify if DB has changed file has rollbacked. If rollback has happened,
we use the next row id present in external storage to update the sqlite
counter. If the next row id is not found on external storage, we use
billion as the starting number. On every insert, we update next row
id in xattr to new_id+BACKUP_FREQUENCY when new_id is greater than
or equal to NEXT_ROW_ID.

Bug: 216743720
Test: Tested manually

Change-Id: Idbbfd11932a7682470527d1cc455ea976011145b