Fix unnecessary redaction for videos in DCIM/Camera
The MediaMetadataRetriever (MMR) attempts to convert any legacy fds
obtained from a modern fd back to a modern fd. Access control is
implemented in the FuseDaemon by checking if the caller has an fd from
the FUSE fs and if so, it simply re-opens the associated file path and
returns the original (modern) fd to the caller.

The previous implementation was conservative and always opened a
redacted fd if an equivalent fd was found on the FUSE fs.
This breaks MMR location metadata access because location will always
be redacted when the data source is a file path or fd because the FUSE
fs will have an equivalent fd by the time convertToModernFd is called.

Now, we retrieve the redaction state of the equivalent opened FUSE fd
and use that to open the returned modern fd.

Test: atest TranscodeTest && atest fuse_node_test
Bug: 205749245

Change-Id: I45670dff4e5348a3b07bf423cd1465c328ad79ea
Merged-In: I45670dff4e5348a3b07bf423cd1465c328ad79ea