Refactored the CloudMediaProvider sync API
Refactored the 'version' and 'generation' concept.
Renamed 'version' to 'media collection id' and 'generation_modifed' to
'sync_generation'.

The idea is that a CMP can have several media collections with the
following properties:
1. Each media collection has a unique id with optional account details
2. Each media collection can contain albums with unique ids across all
collections
3. Each media collection contains media items with unique ids across
all collections.
4. There is only one active collection at a time, which is returned
from the onGetMediaCollectionInfo, formerly onGetMediaInfo.
5. A CMP account switch can be implemented as a new media collection.
6. All media/album queries pass the media collection id to the CMP for
to avoid syncing the wrong media if the collection changes (to be
implemented)

Test: atest PickerSyncControllerTest && CloudPhotoPickerTest
Bug: 214028962
Change-Id: I733cb587a09dba76dd8f78f4f2cf0c720a4fc3ce