Add selection parameter to enforceCallingPermission
Selection and selectionArgs are not used when checking to see if caller
has direct write access. It makes to get all rows of which
owner_package_name is caller package name and volume_name is a volume
where the file is stored. It's good to get only a row associated to the
file.

Bug: 218940560
Test: atest MediaProviderTests, manual test
Change-Id: I409fe6cc3773f000f79efa1992ea57d3485a75ef