FUSE-BPF: Obtain BPF fd in FuseDaemon::Start()
When this userspace implementation was developed, it was still unclear
at what point in time the BPF program of FUSE-BPF would have been
available, and opening that at lookup time was the safest option, making
sure that the whole Android would have been initialized.
Now that the BPF program is loaded by the BPF loader service, the FUSE
daemon can rely on that to be already available when the FUSE file
system is being mounted, so move this initialization in fuse::Start().

Test: ScopedStorageDeviceTest
Bug: 214034634
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: If7cd4ad9861549034065aeece5cb7098380ef722