FUSE-BPF: Unify matching of data and obb paths
The operations to perform on a matching Android/data or Android/obb
paths are the same, so there's no need for having two equivalent regular
expressions for the two different cases, which for matching Android/obb
paths would be both executed.
Thus, merge the two regexes in one that matches both the cases to make
the code cleaner and reduce one regex matching.

Bug: 214034634
Test: ScopedStorageDeviceTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: If3d1095e3e96f1c9f02f81e5cd6a12261901195e