Workaround sdcardfs splice read unavailability
With O_DIRECT, sdcardfs fails read requests for 8KB+ buffers because
of its lack of support for the splice_read operation.

We disable splice on all O_DIRECT reads. We could have limited the
workaround to only sdcardfs devices and reads above 8KB, but this is
simpler and reduces fragmentation on devices.

Test: Manaul
Bug: 177250981
Change-Id: If8d1ebe1e1067d193dc924f2aee2e82032d15633