Temporary quick fix to address crashes around updateSpecialFormatColumn
updateSpecialFormatColumn is called in onIdleMaintenance to update the
special format column value for the images that already exist on the
device.
Since this method is run onIdleMaintenance the crashes do not have
direct user impact, however they may have side-effects on other methods
that are run as part of onIdleMaintenance.

This is a temporary safe fix and we will land a proper fix addressing
the issues and re-enabling this method in onIdleMaintenance soon.

Bug: 206604262
Bug: 207297458
Bug: 219894107
Test: atest MediaProviderTests
Change-Id: Ie3639d1b6f86b7dfd4c2e448937860d05682d32c